Adding first part of Omnidirectional Shadow Maps.

+++ OpenGLCourseApp/DirectionalLight.cpp
@@ -5,7 +5,7 @@ DirectionalLight::DirectionalLight() : Light() {
 	lightProj = glm::ortho(-5.0f, 5.0f, -5.0f, 5.0f, 0.1f, 20.0f);
 }
 
-DirectionalLight::DirectionalLight(GLfloat shadowWidth, GLfloat shadowHeight,
+DirectionalLight::DirectionalLight(GLuint shadowWidth, GLuint shadowHeight,
 									GLfloat red, GLfloat green, GLfloat blue,
 									GLfloat aIntensity, GLfloat dIntensity,
 									GLfloat xDir, GLfloat yDir, GLfloat zDir) : Light(shadowWidth, shadowHeight, red, green, blue, aIntensity, dIntensity) {
